+++ package.json
@@ -12,7 +12,8 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-router-dom": "^6.12.1",
-    "react-scripts": "^4.0.3"
+    "react-scripts": "^4.0.3",
+    "swiper": "^9.4.1"
   },
   "devDependencies": {
     "@cypress/webpack-dev-server": "^1.8.4",
