+++ src/components/ProductSlider/ProductSlider.scss
@@ -1,15 +1,21 @@
 .product-slider {
   display: flex;
-  justify-content: space-between;
   align-items: center;
-
-  padding-bottom: 28px;
+  flex-direction: column;
 
   margin: 0 auto;
   width: 1136px;
 
   @media (max-width: 1136px) {
     padding: 0 20px;
+  }
+
+  &__header {
+    display: flex;
+    justify-content: space-between;
+    width: 1136px;
+
+    padding-bottom: 28px;
   }
 
   &__buttons {
@@ -25,6 +31,18 @@
 
       cursor: pointer;
       transition: all 0.3s;
+
+      &--active {
+        background-color: aqua;
+      }
     }
+  }
+
+  &__content {
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    gap: 16px;
+    padding-bottom: 80px;
   }
 }
