+++ src/components/ProductSlider/ProductSlider.tsx
@@ -1,129 +1,19 @@
 import { useState } from 'react';
-import { Recomended } from '../Recomended/Recomended';
 import { Device } from '../../types/Device';
+import { Card } from '../Card';
 
-const prodCardsPlug: Device[] = [{
-  id: '1',
-  category: 'string',
-  phoneId: 'string',
-  name: 'string',
-  fullPrice: 1499.99,
-  price: 1499,
-  screen: 'string',
-  capacity: 'string',
-  color: 'string',
-  ram: 'string',
-  year: 1984,
-  image: 'string',
-},
-{
-  id: '1',
-  category: 'string',
-  phoneId: 'string',
-  name: 'string',
-  fullPrice: 1499.99,
-  price: 1499,
-  screen: 'string',
-  capacity: 'string',
-  color: 'string',
-  ram: 'string',
-  year: 1984,
-  image: 'string',
-},
-{
-  id: '1',
-  category: 'string',
-  phoneId: 'string',
-  name: 'string',
-  fullPrice: 1499.99,
-  price: 1499,
-  screen: 'string',
-  capacity: 'string',
-  color: 'string',
-  ram: 'string',
-  year: 1984,
-  image: 'string',
-},
-{
-  id: '1',
-  category: 'string',
-  phoneId: 'string',
-  name: 'string',
-  fullPrice: 1499.99,
-  price: 1499,
-  screen: 'string',
-  capacity: 'string',
-  color: 'string',
-  ram: 'string',
-  year: 1984,
-  image: 'string',
-},
-// {
-//   id: '1',
-//   category: 'string',
-//   phoneId: 'string',
-//   name: 'string',
-//   fullPrice: 1499.99,
-//   price: 1499,
-//   screen: 'string',
-//   capacity: 'string',
-//   color: 'string',
-//   ram: 'string',
-//   year: 1984,
-//   image: 'string',
-// },
-// {
-//   id: '1',
-//   category: 'string',
-//   phoneId: 'string',
-//   name: 'string',
-//   fullPrice: 1499.99,
-//   price: 1499,
-//   screen: 'string',
-//   capacity: 'string',
-//   color: 'string',
-//   ram: 'string',
-//   year: 1984,
-//   image: 'string',
-// },
-// {
-//   id: '1',
-//   category: 'string',
-//   phoneId: 'string',
-//   name: 'string',
-//   fullPrice: 1499.99,
-//   price: 1499,
-//   screen: 'string',
-//   capacity: 'string',
-//   color: 'string',
-//   ram: 'string',
-//   year: 1984,
-//   image: 'string',
-// },
-// {
-//   id: '1',
-//   category: 'string',
-//   phoneId: 'string',
-//   name: 'string',
-//   fullPrice: 1499.99,
-//   price: 1499,
-//   screen: 'string',
-//   capacity: 'string',
-//   color: 'string',
-//   ram: 'string',
-//   year: 1984,
-//   image: 'string',
-// }
-];
+type Props = {
+  products: Device[]
+};
 
-export const ProductSlider = () => {
+export const ProductSlider: React.FC<Props> = ({ products }) => {
   const [page, setPage] = useState(1);
 
   const rightButton = () => {
     let point = page + 1;
 
-    if (point > prodCardsPlug.length - 4) {
-      point = prodCardsPlug.length - 4;
+    if (point > products.length - 4) {
+      point = products.length - 4;
       if (point < 0) {
         point = 0;
       }
@@ -145,28 +35,34 @@ export const ProductSlider = () => {
   return (
     <div className="container">
       <div className="product-slider">
-        <h2>You may also like</h2>
+        <div className="product-slider__header">
+          <h2>You may also like</h2>
 
-        <div className="product-slider__buttons">
-          <button
-            type="button"
-            className="product-slider__buttons__button"
-            onClick={leftButton}
-          >
-            {'<'}
-          </button>
+          <div className="product-slider__buttons">
+            <button
+              type="button"
+              className="product-slider__buttons__button"
+              onClick={leftButton}
+            >
+              {'<'}
+            </button>
 
-          <button
-            type="button"
-            className="product-slider__buttons__button"
-            onClick={rightButton}
-          >
-            {'>'}
-          </button>
+            <button
+              type="button"
+              className="product-slider__buttons__button"
+              onClick={rightButton}
+            >
+              {'>'}
+            </button>
+          </div>
+        </div>
+
+        <div className="product-slider__content">
+          {products.slice(page, page + 4).map(product => (
+            <Card phone={product} />
+          ))}
         </div>
       </div>
-
-      <Recomended products={prodCardsPlug} />
     </div>
   );
 };
