+++ src/components/Recomended/Recomended.scss
@@ -1,10 +1,12 @@
 .recomended {
   display: flex;
   justify-content: center;
+  flex-direction: column;
 
   &__block {
     display: flex;
     flex-direction: row;
+    justify-content: center;
     gap: 16px;
     padding-bottom: 80px;
   }
