+++ src/components/Recomended/Recomended.tsx
@@ -1,18 +1,124 @@
 import { Device } from '../../types/Device';
-import { Card } from '../Card';
+import { ProductSlider } from '../ProductSlider/ProductSlider';
 
-type Props = {
-  products: Device[];
-};
+const prodCardsPlug: Device[] = [{
+  id: '1',
+  category: 'stringqwqw',
+  phoneId: 'string112',
+  name: '1',
+  fullPrice: 1499.99,
+  price: 1499,
+  screen: 'string',
+  capacity: 'string',
+  color: 'string',
+  ram: 'string',
+  year: 1984,
+  image: 'string',
+},
+{
+  id: '1',
+  category: 'string',
+  phoneId: 'string',
+  name: '2',
+  fullPrice: 1499.99,
+  price: 1499,
+  screen: 'string',
+  capacity: 'string',
+  color: 'string',
+  ram: 'string',
+  year: 1984,
+  image: 'string',
+},
+{
+  id: '1',
+  category: 'string',
+  phoneId: 'string',
+  name: '3',
+  fullPrice: 1499.99,
+  price: 1499,
+  screen: 'string',
+  capacity: 'string',
+  color: 'string',
+  ram: 'string',
+  year: 1984,
+  image: 'string',
+},
+{
+  id: '1',
+  category: 'string',
+  phoneId: 'string',
+  name: '4',
+  fullPrice: 1499.99,
+  price: 1499,
+  screen: 'string',
+  capacity: 'string',
+  color: 'string',
+  ram: 'string',
+  year: 1984,
+  image: 'string',
+},
+{
+  id: '1',
+  category: 'string',
+  phoneId: 'string',
+  name: '5',
+  fullPrice: 1499.99,
+  price: 1499,
+  screen: 'string',
+  capacity: 'string',
+  color: 'string',
+  ram: 'string',
+  year: 1984,
+  image: 'string',
+},
+{
+  id: '1',
+  category: 'string',
+  phoneId: 'string',
+  name: '6',
+  fullPrice: 1499.99,
+  price: 1499,
+  screen: 'string',
+  capacity: 'string',
+  color: 'string',
+  ram: 'string',
+  year: 1984,
+  image: 'string',
+},
+{
+  id: '1',
+  category: 'string',
+  phoneId: 'string',
+  name: '7',
+  fullPrice: 1499.99,
+  price: 1499,
+  screen: 'string',
+  capacity: 'string',
+  color: 'string',
+  ram: 'string',
+  year: 1984,
+  image: 'string',
+},
+{
+  id: '1',
+  category: 'string',
+  phoneId: 'string',
+  name: '8',
+  fullPrice: 1499.99,
+  price: 1499,
+  screen: 'string',
+  capacity: 'string',
+  color: 'string',
+  ram: 'string',
+  year: 1984,
+  image: 'string',
+},
+];
 
-export const Recomended: React.FC<Props> = ({ products }) => {
+export const Recomended = () => {
   return (
     <div className="recomended">
-      <div className="recomended__block">
-        {products.map((card) => (
-          <Card phone={card} />
-        ))}
-      </div>
+      <ProductSlider products={prodCardsPlug} />
     </div>
   );
 };
